Return useful error when there is no valid build.
When we cannot return a valid build object we had been returning an
error with no explanation. This makes it impossible for the proxy to
do anything nice to let the user know why the build could not be
returned and since there are several possible reasons, we develop a
error code system and messages to assist with this task.